# app/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Voting DLT service API in Go using Gin framework.",
        "title": "Fabric Voting API",
        "contact": {
            "name": "Izqalan Nor'Izad",
            "url": "https://github.com/izqalan",
            "email": "izqalan@duck.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/candidate/{electionID}": {
            "get": {
                "description": "Get candidate by electionID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Candidate"
                ],
                "summary": "Get Candidate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Election ID",
                        "name": "electionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidates fetched",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/election/{electionID}": {
            "get": {
                "description": "Get election by electionID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Election"
                ],
                "summary": "Get Election by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Election ID",
                        "name": "electionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Election created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update election by electionID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Election"
                ],
                "summary": "Update Election",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Election ID",
                        "name": "electionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Election updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}